refactor(reader): address review feedback on viewport alignment
- Extract alignedReaderWidthDp to its own file (ReaderViewportAlignment.kt),
  matching the package's file-per-concern convention and co-locating it with
  its unit test.
- Guard against non-finite availDp (an unbounded-width parent would otherwise
  spin floor(Infinity).toInt() == Int.MAX_VALUE on the UI thread).
- Fallback returns floor(availDp), never the raw fractional width, so the
  drift the function fixes can't sneak back via the no-match path.
- Scope the page-coloured gutter background to paginated mode only, so scroll/
  fixed-layout/double-page keep the bare modifier (truly untouched).
- remember() the aligned width on (maxWidth, density) to keep it off the
  recomposition hot path.
